The Half-duplex chatting system
Half-duplex:
a connection is made and the service starts, only one person can type.
The other participant must wait to get a message before he or she is
prompted to enter a message. Once a message is sent, then the sender
must wait for a reply before being allowed to send another message. One
participant will be on the server side, while the other will be on the
client side.