9-10 Home Finances system
Home Finances. Create a home finance manager. Your solution should be
able to manage savings, checking, money market, certificate of deposit
(CD), and similar accounts. Provide a menu-based interface to each
account as well as operations such as deposits, withdrawals, debits, and
credits. An option should be given to a user to remove transactions as
well. The
data should be stored to file when the user quits the application (but
randomly during execution for backup purposes).