Simple Text Editor.
a simple text editor. A file can be created
from scratch or read and displayed into a Text
widget that can be edited by the user. When the user quits
the application (either with the QUIT button or the Quit/
Exit menu option), the user is prompted whether to save
the changes.